`wipe` prunes all volumes (#210)
Before this change the `wipe` command pruned only anonymous volumes (https://docs.docker.com/engine/reference/commandline/volume_prune/). The `--all `argument was added so that the command prune also other volumes.

## Open questions

Would it make sense to add labels to all `streamr-docker-dev` images? We could `--filter` so that we don't prune some volume that is not related to `streamr-docker-dev`.